Finished 1.0.2
This is now feature complete to the original skeleton ray tracer.  the
GUI has added zoom ability, but a lack of pause/resume ability, which
does not fit easily into java's threading without altering the
ray-tracing algorithm itself to interact with it.

It renders the three 1.0.2 build functions just like they appear in the
book. Build functions are classes that implement a build interface and
are assigned by assigning a builder to a concrete implementation.  This
is "cleaner" and more "java-like" than re-implementing the same method
and selectively including it.