Changed architecture slightly to allow automation
added a listener interface, changed the main to implement it.  also
updated cameras and world to use it. and implemented an automation to
render ALL possible build functions and save each result to an image.
this version reduces the size of the images to 100x100 pixels and
reduces the max recursion depth.